DataUnitDatabaseConnectionProvider.java moved to new package to prevent OSGi conflicts
