+++ dataunit-relational-impl/src/main/java/eu/unifiedviews/dataunit/relational/repository/ManagableRelationalRepository.java
@@ -2,7 +2,7 @@ package eu.unifiedviews.dataunit.relational.repository;
 
 import java.sql.SQLException;
 
-import eu.unifiedviews.dataunit.relational.DataUnitDatabaseConnectionProvider;
+import eu.unifiedviews.dataunit.relational.db.DataUnitDatabaseConnectionProvider;
 
 public interface ManagableRelationalRepository {
