+++ dataunit-relational-impl/src/test/java/eu/unifiedviews/dataunit/relational/impl/RelationalDataUnitImplTest.java
@@ -28,8 +28,8 @@ import eu.unifiedviews.commons.rdf.repository.ManagableRepository;
 import eu.unifiedviews.commons.rdf.repository.RDFException;
 import eu.unifiedviews.commons.rdf.repository.RepositoryFactory;
 import eu.unifiedviews.dataunit.DataUnitException;
-import eu.unifiedviews.dataunit.relational.DataUnitDatabaseConnectionProvider;
 import eu.unifiedviews.dataunit.relational.RelationalDataUnit;
+import eu.unifiedviews.dataunit.relational.db.DataUnitDatabaseConnectionProvider;
 import eu.unifiedviews.dataunit.relational.repository.ManagableRelationalRepository;
 import eu.unifiedviews.dataunit.relational.repository.RelationalException;
 import eu.unifiedviews.dataunit.relational.repository.RelationalRepositoryFactory;
